fix: allow provide mdcConfigs to parseMarkdown
